obj: fix a build failure with UBSAN
Returning a value is an implicit cast, but -fsanitize=undefined
disagrees.

Compare 22d83a2a6f8c5d30278a13a64c8e0f36758366cf ...

+++ src/libpmemobj/critnib.c
@@ -206,7 +206,7 @@ path_mask(sh_t shift)
 static inline unsigned
 slice_index(uint64_t key, sh_t shift)
 {
-	return (key >> shift) & NIB;
+	return (uint32_t)((key >> shift) & NIB);
 }
 
 /*
